requirements.txt
pandas
numpy
scikit-learn
tqdm
Pillow
jupyter
matplotlib
ipykernel
torch
torchvision
torchaudio